aitestmock/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aitestmock"
version = "0.1.0"
dependencies = [
    "Flask>=3.1.2",
    "Flask-CORS>=4.0.0",
    "Werkzeug>=3.1.0",
    "requests>=2.32.5",
    "python-dotenv>=1.1.1",
    "gunicorn>=23.0.0"
]

[project.scripts]
start-server = "aitestmock.server:main"

[tool.setuptools.packages.find]
where = ["src"]
